refactor wallet

+++ src/token.rs
@@ -1,13 +1,13 @@
 pub struct Token {
     pub name: String,
-    pub amount: u128,
+    pub amount: f64,
 }
 
 impl Token{
     pub fn new() -> Token{
         Token{
             name: String::from("ETH"),
-            amount: 0
+            amount: 0.0
         }
     }
 }
